feat: Precisely determine if bzlmod is enabled
Rather than relying on just the version number we can use `bazel info
starlark-semantics` as described in https://github.com/bazelbuild/bazel/issues/21303
